Directive edits

+++ Angular_JS/angular-skeleton-master/app/common/directives/employeeDelete.js
@@ -1,14 +1,21 @@
 angular.module('employee.delete', ['home.service'])
 .directive('employeeDelete',['$location','homeService', function($location,homeService) {
   return {
-   restrict: 'E',
-   templateUrl: 'app/home/views/delete_emp.html',
-   link: function (scope, elem, attr) {
-    elem.on("click",function(){
-      homeService.deleteEmployee(attr.email);
-        //$rootScope.employees = response;
-        $location.path('/home');
-     });
-    }
+    restrict: 'E',
+    templateUrl: 'app/home/views/delete_emp.html',
+    link: function (scope, $elem, attr) {
+      $elem.on("click",function(){
+        var r = confirm("Are you sure want to delete?");
+          if(r == true){
+            var response = homeService.deleteEmployee(attr.email);
+            console.log("hi its directive");
+            console.log(response);
+            $location.path('/home');
+          }
+          else{
+            alert("No records are deleted...!!")
+          }
+      });
+    },
   };
 }]);
